add query and change file
